fix(trajectory-ingest): cast Cell_Size to Python float
UDFManager.put silently writes 0 when handed numpy float32 (the dtype
MDAnalysis returns for universe.dimensions). The previous code passed
`dims[0] / 10.0` directly to TopExporter._append_structure, which in
turn called `uobj.put(cell[0], "...Cell_Size.a", "[nm]")` and silently
recorded 0 instead of the actual box length.

Downstream effect: trajectory.udf had Cell_Size = [0, 0, 0, 90, 90, 90],
which made abmptools.udf_io.moveintocell hang in an infinite while-loop
(`while centerOfMol[j] > cell[j]: centerOfMol[j] -= cell[j]` is a no-op
when cell[j] == 0). fcews-manybody's setupgetcontact_gromacs path then
appeared to "time out" but was actually stuck — 4_getpdb never produced
*.ajf for a propane-butane smoke run after 47 min of wall.

Confirmed on the propane-butane Fugaku smoke run: with the fix
trajectory.udf gets Cell_Size = [32.37, 32.37, 32.37, 90, 90, 90] Å
and moveintocell completes in 0.13 s (was: timeout at 47 min).

Fixes applied at three levels for defence in depth:
1. trajectory_ingest.frames_from_xtc — cast cell to float at the source
2. top_exporter._append_structure — float() in the put call
3. gro2udf.udf_writer (static GRO path) — same defensive cast

+++ abmptools/amorphous/trajectory_ingest.py
@@ -184,9 +184,17 @@ def frames_from_xtc(
                 )
         # MDAnalysis positions are Å — convert to nm (GROMACS convention)
         coords_nm = (universe.atoms.positions / 10.0).tolist()
-        # universe.dimensions: [a, b, c, alpha, beta, gamma] in Å / deg
+        # universe.dimensions: [a, b, c, alpha, beta, gamma] in Å / deg.
+        # Cast each element to Python float — UDFManager.put silently
+        # writes 0 when handed numpy float32 (the MDAnalysis dtype),
+        # which downstream produces a trajectory.udf with
+        # Cell_Size = [0, 0, 0, ...] and makes
+        # abmptools.udf_io.moveintocell hang in an infinite while-loop
+        # (`centerOfMol[j] -= cell[j]` with cell[j]=0).
         dims = universe.dimensions
-        cell_nm = [dims[0] / 10.0, dims[1] / 10.0, dims[2] / 10.0]
+        cell_nm = [float(dims[0] / 10.0),
+                   float(dims[1] / 10.0),
+                   float(dims[2] / 10.0)]
         frames.append(
             GROFrameData(
                 step=i,
